+++ tests/next_tests/unit_tests/iterator_tests/test_type_inference.py
@@ -15,8 +15,7 @@
 import numpy as np
 
 from gt4py.next.common import Dimension
-from gt4py.next.ffront import itir_makers as im
-from gt4py.next.iterator import ir, type_inference as ti
+from gt4py.next.iterator import ir, makers as im, type_inference as ti
 from gt4py.next.iterator.embedded import NeighborTableOffsetProvider
 from gt4py.next.iterator.runtime import CartesianAxis
