+++ tests/next_tests/unit_tests/iterator_tests/transforms_tests/test_inline_lifts.py
@@ -14,8 +14,7 @@
 
 import pytest
 
-from gt4py.next.ffront import itir_makers as im
-from gt4py.next.iterator import ir
+from gt4py.next.iterator import makers as im
 from gt4py.next.iterator.transforms.inline_lifts import InlineLifts
